requirements.txt
# Story to Image Generator - Requirements
# Version: 3.0.0

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data Validation & Models
pydantic>=2.5.0

# HTTP Requests (Sync & Async)
requests>=2.31.0
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# File Handling
pathlib2>=2.3.7

# Development & Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing async endpoints

# Monitoring & Logging (Optional but recommended)
structlog>=23.2.0

# Streamlit Frontend Requirements
# Story to Image Generator v3.0

streamlit>=1.29.0
requests>=2.31.0
Pillow>=10.0.0
aiohttp>=3.8.0

# Optional: For enhanced functionality
plotly>=5.15.0  # For advanced charts/visualizations
pandas>=2.0.0   # For data handling (if needed)
numpy>=1.24.0   # For image processing utilities